Improve quantization-related issues
This patch improves situations where the collision detection thinks that
the hash is some value, but saving the image and hashing the saved image
results in a slightly different value. One detail that was contributing
to this was that the optimization represents images using 32-bit floats
per channel, whereas the saved image is quantized to 8 bits per channel.
This patch fixes that issue by computing the forward _and backward pass_
with respect to the quantized image (computing derivatives with respect
to the quantized version of the image is fine; this can be thought of as
BPDA; see https://arxiv.org/abs/1802.00420).

The adversarial examples that are produced by the optimization are still
not necessarily "robust"; they are not necessarily likely to survive
mutations like noise or lossy compression (e.g. if saved as a JPEG). A
future patch could use EOT (https://arxiv.org/abs/1707.07397) to find
more robust collisions.

This patch also fixes a bug where the image was saved _after_ the
gradient descent step, so the hash of the saved image was not
necessarily the hash that was computed during the forward pass.